+ Added mapTableGetMap(). + Added mapTableContainerFromData().

+++ llc/gpk_block.cpp
@@ -50,7 +50,7 @@
 	return 0;
 }
 
-::gpk::error_t								gpk::blockFilePath			(::gpk::array_pod<char_t> & finalPath, const ::gpk::view_const_char & dbName, const ::gpk::view_const_char & dbPath, const uint32_t containers, const uint8_t indexContainer) {
+::gpk::error_t								gpk::blockFilePath			(::gpk::array_pod<char_t> & finalPath, const ::gpk::view_const_char & dbName, const ::gpk::view_const_char & dbPath, const uint32_t containers, const uint32_t indexContainer) {
 	finalPath									= dbPath;
 	if(finalPath.size())
 		finalPath.push_back('/');
